refactor: Reorganise codebase into packages
