+++ src/main/java/clanker/task/EventTask.java
@@ -1,3 +1,7 @@
+package clanker.task;
+
+import fmt.Formatters;
+
 import java.time.LocalDateTime;
 
 public class EventTask extends Task {
