+++ src/main/java/fmt/CommandParser.java
@@ -1,3 +1,5 @@
+package fmt;
+
 import java.util.Arrays;
 import java.util.HashMap;
