+++ src/main/java/fmt/Formatters.java
@@ -1,3 +1,5 @@
+package fmt;
+
 import java.time.format.DateTimeFormatter;
 
 public class Formatters {
