+++ src/main/java/serde/Serialisable.java
@@ -1,3 +1,5 @@
+package serde;
+
 public interface Serialisable {
     String serialise();
 }
